% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAAD.R
\docType{data}
\name{GAAD}
\alias{GAAD}
\title{Current-status GAAD data set}
\format{
A data frame with 32735 rows and 9 variables:
\describe{
\item{id}{Subject ID}
\item{site}{tooth-site, for each subject}
\item{Female}{Gender, 0: Male, 1: Female}
\item{BMI}{Body Mass Index, for each subject}
\item{Smoker}{Smoking status, 0: non-smoker, 1: smoker}
\item{Hba1c}{Glycemic Level, 0: uncontrolled, 1: high glycemic level}
\item{Maxilla}{0: tooth-site in lower jaw, 1: tooth-site in upper jaw}
\item{CSTime}{The current status (inspection) time of the (adult) tooth, in years.}
\item{state}{State as defined by clinical attachment level (CAL), occupied when inspected. States 1, 2, 3 and 4 correspond to 'Healthy', 'Mild', 'Moderate' and 'Severe' level of periodontal disease. State 4 is the absorbing state}
...
}
}
\usage{
GAAD
}
\description{
Data from the cross-sectional study of 288 Gullah-speaking African-American diabetics (GAAD) where their
periodontal disease (PD) status was evaluated at a given inspection time. Incidence of periodontal disease
was determined by measuring the clinical attachment level at six sites of each tooth, excluding molars,
in the subjects' mouth. The data contains the current status times, states that describe severity of PD,
and clinical/demographic information.
}
\references{
{
\insertRef{fernandes2009}{mspack2}
}
}
\keyword{datasets}
