% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOP.R
\name{SOP}
\alias{SOP}
\title{Nonparametric estimators for Current-status data from multistate models}
\usage{
SOP(
  data.type,
  dat,
  tree,
  start.probs,
  weight = "none",
  bw = NULL,
  ngrid = 1000,
  pavY = TRUE,
  cond = "none",
  Zval = NULL,
  cutoffs = NULL
)
}
\arguments{
\item{data.type}{a character string specifying the structure of the observed data. This can either be 'uncorrelated' or 'cluster-correlated'.}

\item{dat}{a data frame with columns 'id', 'time', 'state' indicating unique subject ID, inspection time and state occupation at inspection time. If data.type='cluster-correlated',
the data frame should also include 'cID' and 'csize' which are variables that specifies the unique cluster ID and cluster size.}

\item{tree}{a graphNEL graph with the nodes and edges of the multistate model.}

\item{start.probs}{a named numeric vector specifying the initial occupation probabilities of each state. The names of the elements must correspond to the nodes of tree.}

\item{weight}{a character vector to specify the weight for each subject if data.type='cluster-correlated'. Weights can be either 'ICS' or 'none'.}

\item{bw}{a numeric value specifying a bandwidth sequence for the inspection times. Default is NULL, where the bandwidth sequence is chosen by the criteria of \insertCite{wand1994;textual}{mspack2}.}

\item{ngrid}{an integer specifying specifying the number of grid points from 0 to max of the inspection times. Default is 1000.}

\item{pavY}{a logical argument specifying whether the pooled adjascent violator algorithms should be used to estimate the at-risk set. Default is TRUE.}

\item{cond}{a character argument specifying a single continuous covariate for conditional estimators. This is based on the method by \insertCite{lan2017;textual}{mspack2}. Default is 'none' which indicates marginal estimators are of interest.}

\item{Zval}{a numeric value indicating the value of the covariate at which the conditional estimator is evaluated at. Default is NULL where the conditional estimator is evaluated at the first quartile of the single covariate.}

\item{cutoffs}{a numeric vector specifying the time points at which the estimates are returned. Default is NULL, which returns the SOPs at all inspection time points.}
}
\value{
a data frame of state occupation probabilities of each state in the multistate model.
}
\description{
Nonparametric estimators for Current-status data from multistate models
}
\details{
Utilizes theory on nonparametric regression to estimate the counting process and at-risk sets for current status data from
general multistate models. The counting process and at-risk sets are necessary functionals for estimating the state occupation
probabilities.
}
\examples{
## Three-state tracking model
Nodes <- c('1','2','3')
Edges <- list('1'=list(edges=c('2')), '2'=list(edges=c('3')),'3'=list(edges=NULL))

## Constructs the tree
tree <- new('graphNEL', nodes=Nodes, edgeL=Edges, edgemode='directed')

## simple data simulation
m <- 5
cdat <- lapply(1:m, function(x){
  ni <- rpois(n=1, lambda=10)+2 # cluster size
  cID <- rep(x=x, times=ni) # cluster ids
  time <- rweibull(n=ni, shape=3, scale=5) # inspection time
  state <- sample(x=c(as.numeric(Nodes)), size=ni, replace=TRUE) # state at inspection time
  csize <- rep(x=ni, times=ni) # cluster sizes
  data.frame(cID=cID, time=time, state=state, csize=csize)
})
cdat <- do.call(rbind, cdat)
cdat$id <- 1:nrow(cdat)

## initial probabilities - under the assumption that everyone states in state 1
start.probs <- c(1, 0, 0)  # assumes all subjects start from state 1 and time 0
names(start.probs) <- Nodes

## Estimate the state occupation probabilities
SOP(data.type='cluster-correlated',dat=cdat, tree=tree, start.probs=start.probs,
    ngrid=1000, weight='none', pavY=TRUE, cond='none', Zval=NULL,
    cutoffs = seq(from=min(cdat$time), to=max(cdat$time), length.out=10))
}
\references{
{
\insertRef{datta2006}{mspack2}
}
}
\author{
Samuel Anyaso-Samuel, Dipankar Bandyopadhyay, Somnath Datta
}
